==> notebooks/step_2_pandas_profiling_prep_data.ipynb <==
# Objective

Execute a quick profiling on the data to give us some general idea of the instance and price distribution.

# Code

## Load libs

import pandas as pd
from pandas_profiling import ProfileReport

## Input params

interim_dir = '../data/interim'
in_fname = 'step_1_aws_filtered_sample.csv.zip'
compression = 'zip'

report_dir = '../reports/pandas_profiling'
out_fname = 'pandas_profile_aws_pricing.html'

# Papermill parameters injection ... do not delete!

## Load data

data = pd.read_csv(f'{interim_dir}/{in_fname}', 
                   parse_dates=['Timestamp'],
                   compression=compression, 
                   index_col=0)

print(data.dtypes)
print(data.shape)
data.head()

## Load Pandas Profiler

Execute pandas profiler and place it in the report folder

report = ProfileReport(data)
report.to_file(f'{report_dir}/{out_fname}')
